Remove AppFeature+Navigation.swift and move Destination enum to AppFeature.swift
